rearranged more download code files
